% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fsx_package.R
\docType{package}
\name{paws.fsx-package}
\alias{paws.fsx}
\alias{paws.fsx-package}
\title{paws.fsx: Amazon FSx}
\description{
Amazon FSx is a fully managed service that makes it
    easy for storage and application administrators to launch and use
    shared file storage.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
