% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fsx_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Tags an Amazon FSx resource}
\usage{
tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the Amazon FSx resource that you want to tag.}

\item{Tags}{[required] A list of tags for the resource. If a tag with a given key already exists, the value is replaced by the one specified in this parameter.}
}
\description{
Tags an Amazon FSx resource.
}
\section{Accepted Parameters}{
\preformatted{tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
# This operation tags an Amazon FSx resource.
\donttest{tag_resource(
  ResourceARN = "arn:aws:fsx:us-east-1:012345678912:file-system/fs-0498eed5fe91001ec",
  Tags = list(
    list(
      Key = "Name",
      Value = "MyFileSystem"
    )
  )
)}

}
